pyproject.toml
[project]
name = "myvenv"
version = "0.1.0"
description = ""
authors = [
    {name = "Manal Farouqi",email = "farouqimanal@gmail.com"}
]
readme = "README.md"
requires-python = "3.11.14"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "joblib (>=1.5.2,<2.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "keras (>=3.12.0,<4.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "dlib (>=20.0.0,<21.0.0)",
    "imutils (>=0.5.4,<0.6.0)",
    "fastapi[standard] (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"